feat: add stable flag to PoppunkWrapper initialization
This commit adds the stable flag to the initialization of the PoppunkWrapper class in the beebop/poppunkWrapper.py file. The stable flag is passed as an argument to the PoppunkAssigner constructor, allowing for stable assignments of clusters. This change improves the stability of the clustering process.

Refactor args.json

This commit updates the args.json file in the beebop/resources directory. It adds a new "stable" flag with a value of "false". This flag is used to control the stability of the clustering process in PoppunkWrapper. By default, the clustering process is not stable.